+++ src/migration.rs
@@ -10,7 +10,7 @@ use std::{
     net::IpAddr,
     path::{Path, PathBuf},
 };
-use tracing::info;
+use tracing::{info, warn};
 
 /// The range of versions that use the current database format.
 ///
@@ -32,7 +32,7 @@ use tracing::info;
 /// // the database format won't be changed in the future alpha or beta versions.
 /// const COMPATIBLE_VERSION: &str = ">=0.5.0-alpha.2,<=0.5.0-alpha.4";
 /// ```
-const COMPATIBLE_VERSION_REQ: &str = ">=0.25.0,<=0.26.0-alpha.1";
+const COMPATIBLE_VERSION_REQ: &str = ">0.25.0,<=0.26.0-alpha.2";
 
 /// Migrates data exists in `PostgresQL` to Rocksdb if necessary.
 ///
@@ -184,6 +184,11 @@ pub fn migrate_data_dir<P: AsRef<Path>>(data_dir: P, backup_dir: P) -> Result<()
             Version::parse("0.25.0")?,
             migrate_0_24_to_0_25,
         ),
+        (
+            VersionReq::parse(">=0.25.0,<0.26.0")?,
+            Version::parse("0.26.0")?,
+            migrate_0_25_to_0_26,
+        ),
     ];
 
     let mut store = super::Store::new(data_dir, backup_dir)?;
@@ -257,6 +262,171 @@ fn read_version_file(path: &Path) -> Result<Version> {
         .read_to_string(&mut ver)
         .context("cannot read VERSION")?;
     Version::parse(&ver).context("cannot parse VERSION")
+}
+
+fn migrate_0_25_to_0_26(store: &super::Store) -> Result<()> {
+    use crate::collections::Indexed;
+    use crate::node::{Node, NodeSetting};
+    use crate::IterableMap;
+    use bincode::Options;
+    use chrono::{DateTime, Utc};
+    use std::collections::HashMap;
+
+    type PortNumber = u16;
+
+    #[derive(Deserialize, Serialize)]
+    pub struct OldNode {
+        pub id: u32,
+        pub creation_time: DateTime<Utc>,
+        as_is: Option<OldNodeSetting>,
+        to_be: Option<OldNodeSetting>,
+    }
+
+    #[allow(clippy::struct_excessive_bools, clippy::module_name_repetitions)]
+    #[derive(Deserialize, Serialize)]
+    struct OldNodeSetting {
+        pub name: String,
+        pub customer_id: u32,
+        pub description: String,
+        pub hostname: String,
+
+        pub review: bool,
+        pub review_port: Option<PortNumber>,
+        pub review_web_port: Option<PortNumber>,
+
+        pub piglet: bool,
+        pub piglet_giganto_ip: Option<IpAddr>,
+        pub piglet_giganto_port: Option<PortNumber>,
+        pub piglet_review_ip: Option<IpAddr>,
+        pub piglet_review_port: Option<PortNumber>,
+        pub save_packets: bool,
+        pub http: bool,
+        pub office: bool,
+        pub exe: bool,
+        pub pdf: bool,
+        pub html: bool,
+        pub txt: bool,
+        pub smtp_eml: bool,
+        pub ftp: bool,
+
+        pub giganto: bool,
+        pub giganto_ingestion_ip: Option<IpAddr>,
+        pub giganto_ingestion_port: Option<PortNumber>,
+        pub giganto_publish_ip: Option<IpAddr>,
+        pub giganto_publish_port: Option<PortNumber>,
+        pub giganto_graphql_ip: Option<IpAddr>,
+        pub giganto_graphql_port: Option<PortNumber>,
+        pub retention_period: Option<u16>,
+
+        pub reconverge: bool,
+        pub reconverge_review_ip: Option<IpAddr>,
+        pub reconverge_review_port: Option<PortNumber>,
+        pub reconverge_giganto_ip: Option<IpAddr>,
+        pub reconverge_giganto_port: Option<PortNumber>,
+
+        pub hog: bool,
+        pub hog_review_ip: Option<IpAddr>,
+        pub hog_review_port: Option<PortNumber>,
+        pub hog_giganto_ip: Option<IpAddr>,
+        pub hog_giganto_port: Option<PortNumber>,
+        pub protocols: bool,
+        pub protocol_list: HashMap<String, bool>,
+
+        pub sensors: bool,
+        pub sensor_list: HashMap<String, bool>,
+    }
+
+    impl From<OldNodeSetting> for NodeSetting {
+        fn from(input: OldNodeSetting) -> Self {
+            Self {
+                customer_id: input.customer_id,
+                description: input.description,
+                hostname: input.hostname,
+                review: input.review,
+                review_port: input.review_port,
+                review_web_port: input.review_web_port,
+                piglet: input.piglet,
+                piglet_giganto_ip: input.piglet_giganto_ip,
+                piglet_giganto_port: input.piglet_giganto_port,
+                piglet_review_ip: input.piglet_review_ip,
+                piglet_review_port: input.piglet_review_port,
+                save_packets: input.save_packets,
+                http: input.http,
+                office: input.office,
+                exe: input.exe,
+                pdf: input.pdf,
+                html: input.html,
+                txt: input.txt,
+                smtp_eml: input.smtp_eml,
+                ftp: input.ftp,
+                giganto: input.giganto,
+                giganto_ingestion_ip: input.giganto_ingestion_ip,
+                giganto_ingestion_port: input.giganto_ingestion_port,
+                giganto_publish_ip: input.giganto_publish_ip,
+                giganto_publish_port: input.giganto_publish_port,
+                giganto_graphql_ip: input.giganto_graphql_ip,
+                giganto_graphql_port: input.giganto_graphql_port,
+                retention_period: input.retention_period,
+                reconverge: input.reconverge,
+                reconverge_review_ip: input.reconverge_review_ip,
+                reconverge_review_port: input.reconverge_review_port,
+                reconverge_giganto_ip: input.reconverge_giganto_ip,
+                reconverge_giganto_port: input.reconverge_giganto_port,
+                hog: input.hog,
+                hog_review_ip: input.hog_review_ip,
+                hog_review_port: input.hog_review_port,
+                hog_giganto_ip: input.hog_giganto_ip,
+                hog_giganto_port: input.hog_giganto_port,
+                protocols: input.protocols,
+                protocol_list: input.protocol_list,
+                sensors: input.sensors,
+                sensor_list: input.sensor_list,
+            }
+        }
+    }
+
+    impl TryFrom<OldNode> for Node {
+        type Error = &'static str;
+
+        fn try_from(input: OldNode) -> Result<Self, Self::Error> {
+            let (name, name_draft) = match (&input.as_is, &input.to_be) {
+                (None, None) => {
+                    return Err("Both `as_is` and `to_be` are `None`");
+                }
+                (None, Some(to_be)) => (to_be.name.to_string(), None),
+                (Some(as_is), None) => (as_is.name.to_string(), None),
+                (Some(as_is), Some(to_be)) => {
+                    (as_is.name.to_string(), Some(to_be.name.to_string()))
+                }
+            };
+
+            Ok(Self {
+                id: input.id,
+                name,
+                name_draft,
+                setting: input.as_is.map(std::convert::Into::into),
+                setting_draft: input.to_be.map(std::convert::Into::into),
+                creation_time: input.creation_time,
+            })
+        }
+    }
+
+    let node_db = store.node_map();
+    for (_key, old_value) in node_db.iter_forward()? {
+        let old_node = bincode::DefaultOptions::new()
+            .deserialize::<OldNode>(&old_value)
+            .context("Failed to migrate node database: invalid node value")?;
+
+        match TryInto::<Node>::try_into(old_node) {
+            Ok(new_node) => {
+                node_db.overwrite(&new_node)?;
+            }
+            Err(e) => {
+                warn!("Skip the migration for an item: {e}");
+            }
+        }
+    }
+    Ok(())
 }
 
 fn migrate_0_24_to_0_25(store: &super::Store) -> Result<()> {
@@ -371,10 +541,11 @@ fn migrate_filter_map(store: &super::Store) -> Result<()> {
 
 fn migrate_node(store: &super::Store) -> Result<()> {
     use crate::collections::Indexed;
-    use crate::node::{Node, NodeSetting};
-    use crate::IterableMap;
+    use crate::types::FromKeyValue;
+    use crate::{Indexable, IterableMap};
     use bincode::Options;
     use chrono::{DateTime, Utc};
+    use std::borrow::Cow;
     use std::collections::HashMap;
 
     type PortNumber = u16;
@@ -439,13 +610,110 @@ fn migrate_node(store: &super::Store) -> Result<()> {
         pub apply_in_progress: bool,
     }
 
-    impl From<OldNode> for Node {
+    #[allow(clippy::struct_excessive_bools, clippy::module_name_repetitions)]
+    #[derive(Deserialize, Serialize)]
+    pub struct NewNodeSetting {
+        pub name: String,
+        pub customer_id: u32,
+        pub description: String,
+        pub hostname: String,
+
+        pub review: bool,
+        pub review_port: Option<PortNumber>,
+        pub review_web_port: Option<PortNumber>,
+
+        pub piglet: bool,
+        pub piglet_giganto_ip: Option<IpAddr>,
+        pub piglet_giganto_port: Option<PortNumber>,
+        pub piglet_review_ip: Option<IpAddr>,
+        pub piglet_review_port: Option<PortNumber>,
+        pub save_packets: bool,
+        pub http: bool,
+        pub office: bool,
+        pub exe: bool,
+        pub pdf: bool,
+        pub html: bool,
+        pub txt: bool,
+        pub smtp_eml: bool,
+        pub ftp: bool,
+
+        pub giganto: bool,
+        pub giganto_ingestion_ip: Option<IpAddr>,
+        pub giganto_ingestion_port: Option<PortNumber>,
+        pub giganto_publish_ip: Option<IpAddr>,
+        pub giganto_publish_port: Option<PortNumber>,
+        pub giganto_graphql_ip: Option<IpAddr>,
+        pub giganto_graphql_port: Option<PortNumber>,
+        pub retention_period: Option<u16>,
+
+        pub reconverge: bool,
+        pub reconverge_review_ip: Option<IpAddr>,
+        pub reconverge_review_port: Option<PortNumber>,
+        pub reconverge_giganto_ip: Option<IpAddr>,
+        pub reconverge_giganto_port: Option<PortNumber>,
+
+        pub hog: bool,
+        pub hog_review_ip: Option<IpAddr>,
+        pub hog_review_port: Option<PortNumber>,
+        pub hog_giganto_ip: Option<IpAddr>,
+        pub hog_giganto_port: Option<PortNumber>,
+        pub protocols: bool,
+        pub protocol_list: HashMap<String, bool>,
+
+        pub sensors: bool,
+        pub sensor_list: HashMap<String, bool>,
+    }
+    #[derive(Deserialize, Serialize)]
+    pub struct NewNode {
+        pub id: u32,
+        pub creation_time: DateTime<Utc>,
+        pub as_is: Option<NewNodeSetting>,
+        pub to_be: Option<NewNodeSetting>,
+    }
+
+    impl FromKeyValue for NewNode {
+        fn from_key_value(_key: &[u8], value: &[u8]) -> anyhow::Result<Self> {
+            Ok(bincode::DefaultOptions::new().deserialize(value)?)
+        }
+    }
+
+    impl Indexable for NewNode {
+        fn key(&self) -> Cow<[u8]> {
+            if let Some(as_is) = &self.as_is {
+                Cow::from(as_is.name.as_bytes())
+            } else if let Some(to_be) = &self.to_be {
+                Cow::from(to_be.name.as_bytes())
+            } else {
+                panic!("Both `as_is` and `to_be` are `None`");
+            }
+        }
+
+        fn value(&self) -> Vec<u8> {
+            bincode::DefaultOptions::new()
+                .serialize(self)
+                .expect("serializable")
+        }
+
+        fn set_index(&mut self, index: u32) {
+            self.id = index;
+        }
+
+        fn index(&self) -> u32 {
+            self.id
+        }
+
+        fn make_indexed_key(key: Cow<[u8]>, _index: u32) -> Cow<[u8]> {
+            key
+        }
+    }
+
+    impl From<OldNode> for NewNode {
         fn from(input: OldNode) -> Self {
             Self {
                 id: input.id,
                 creation_time: input.creation_time,
                 as_is: None,
-                to_be: Some(NodeSetting {
+                to_be: Some(NewNodeSetting {
                     name: input.name,
                     customer_id: input.customer_id,
                     description: input.description,
@@ -499,7 +767,7 @@ fn migrate_node(store: &super::Store) -> Result<()> {
         let old_node = bincode::DefaultOptions::new()
             .deserialize::<OldNode>(&old_value)
             .context("Failed to migrate node database: invalid node value")?;
-        let new_node: Node = old_node.into();
+        let new_node: NewNode = old_node.into();
         node_db.overwrite(&new_node)?;
     }
     Ok(())
@@ -944,6 +1212,180 @@ mod tests {
             .flat_map(|u| filters.iter().map(|f| (u.to_string(), f.to_string())))
             .collect();
         assert_eq!(res, testers);
+    }
+
+    #[test]
+    fn migrate_0_25_to_0_26_node() {
+        type PortNumber = u16;
+        use crate::node::Node;
+        use crate::{collections::Indexed, Indexable};
+        use bincode::Options;
+        use chrono::{DateTime, Utc};
+        use serde::{Deserialize, Serialize};
+        use std::{
+            collections::HashMap,
+            net::{IpAddr, Ipv4Addr},
+        };
+
+        #[derive(Deserialize, Serialize, Clone)]
+        pub struct OldNode {
+            pub id: u32,
+            pub creation_time: DateTime<Utc>,
+            pub as_is: Option<OldNodeSetting>,
+            pub to_be: Option<OldNodeSetting>,
+        }
+
+        #[allow(clippy::struct_excessive_bools, clippy::module_name_repetitions)]
+        #[derive(Deserialize, Serialize, Clone)]
+        struct OldNodeSetting {
+            pub name: String,
+            pub customer_id: u32,
+            pub description: String,
+            pub hostname: String,
+
+            pub review: bool,
+            pub review_port: Option<PortNumber>,
+            pub review_web_port: Option<PortNumber>,
+
+            pub piglet: bool,
+            pub piglet_giganto_ip: Option<IpAddr>,
+            pub piglet_giganto_port: Option<PortNumber>,
+            pub piglet_review_ip: Option<IpAddr>,
+            pub piglet_review_port: Option<PortNumber>,
+            pub save_packets: bool,
+            pub http: bool,
+            pub office: bool,
+            pub exe: bool,
+            pub pdf: bool,
+            pub html: bool,
+            pub txt: bool,
+            pub smtp_eml: bool,
+            pub ftp: bool,
+
+            pub giganto: bool,
+            pub giganto_ingestion_ip: Option<IpAddr>,
+            pub giganto_ingestion_port: Option<PortNumber>,
+            pub giganto_publish_ip: Option<IpAddr>,
+            pub giganto_publish_port: Option<PortNumber>,
+            pub giganto_graphql_ip: Option<IpAddr>,
+            pub giganto_graphql_port: Option<PortNumber>,
+            pub retention_period: Option<u16>,
+
+            pub reconverge: bool,
+            pub reconverge_review_ip: Option<IpAddr>,
+            pub reconverge_review_port: Option<PortNumber>,
+            pub reconverge_giganto_ip: Option<IpAddr>,
+            pub reconverge_giganto_port: Option<PortNumber>,
+
+            pub hog: bool,
+            pub hog_review_ip: Option<IpAddr>,
+            pub hog_review_port: Option<PortNumber>,
+            pub hog_giganto_ip: Option<IpAddr>,
+            pub hog_giganto_port: Option<PortNumber>,
+            pub protocols: bool,
+            pub protocol_list: HashMap<String, bool>,
+
+            pub sensors: bool,
+            pub sensor_list: HashMap<String, bool>,
+        }
+
+        impl Indexable for OldNode {
+            fn key(&self) -> Cow<[u8]> {
+                if let Some(as_is) = &self.as_is {
+                    Cow::from(as_is.name.as_bytes())
+                } else if let Some(to_be) = &self.to_be {
+                    Cow::from(to_be.name.as_bytes())
+                } else {
+                    panic!("Both `as_is` and `to_be` are `None`");
+                }
+            }
+
+            fn value(&self) -> Vec<u8> {
+                bincode::DefaultOptions::new()
+                    .serialize(self)
+                    .expect("serializable")
+            }
+
+            fn set_index(&mut self, index: u32) {
+                self.id = index;
+            }
+
+            fn make_indexed_key(key: Cow<[u8]>, _index: u32) -> Cow<[u8]> {
+                key
+            }
+
+            fn index(&self) -> u32 {
+                self.id
+            }
+        }
+
+        let settings = TestSchema::new();
+        let node_db = settings.store.node_map();
+
+        let old_node = OldNode {
+            id: 0,
+            creation_time: Utc::now(),
+            as_is: None,
+            to_be: Some(OldNodeSetting {
+                name: "name".to_string(),
+                customer_id: 20,
+                description: "description".to_string(),
+                hostname: "host".to_string(),
+                review: true,
+                review_port: Some(4040),
+                review_web_port: Some(8442),
+                piglet: true,
+                piglet_giganto_ip: Some(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 2))),
+                piglet_giganto_port: Some(3030),
+                piglet_review_ip: Some(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1))),
+                piglet_review_port: Some(4040),
+                save_packets: true,
+                http: false,
+                office: false,
+                exe: false,
+                pdf: false,
+                html: false,
+                txt: false,
+                smtp_eml: false,
+                ftp: false,
+                giganto: true,
+                giganto_ingestion_ip: Some(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
+                giganto_ingestion_port: Some(3030),
+                giganto_publish_ip: Some(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
+                giganto_publish_port: Some(3050),
+                giganto_graphql_ip: Some(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0))),
+                giganto_graphql_port: Some(5050),
+                retention_period: Some(100),
+                reconverge: false,
+                reconverge_review_ip: Some(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1))),
+                reconverge_review_port: Some(4040),
+                reconverge_giganto_ip: Some(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1))),
+                reconverge_giganto_port: Some(3050),
+                hog: true,
+                hog_review_ip: Some(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1))),
+                hog_review_port: Some(4040),
+                hog_giganto_ip: Some(IpAddr::V4(Ipv4Addr::new(1, 1, 1, 1))),
+                hog_giganto_port: Some(3050),
+                protocols: false,
+                protocol_list: HashMap::new(),
+                sensors: false,
+                sensor_list: HashMap::new(),
+            }),
+        };
+
+        assert!(node_db.insert(old_node.clone()).is_ok());
+        let (db_dir, backup_dir) = settings.close();
+        let settings = TestSchema::new_with_dir(db_dir, backup_dir);
+        assert!(super::migrate_0_25_to_0_26(&settings.store).is_ok());
+
+        let node_db = settings.store.node_map();
+        let new_node = node_db.get_by_key("name".as_bytes()).unwrap().unwrap();
+        let new_node: Node = bincode::DefaultOptions::new()
+            .deserialize(new_node.as_ref())
+            .expect("deserializable");
+
+        assert_eq!(new_node.id, 0);
+        assert_eq!(new_node.name, "name");
     }
 
     #[test]